rename connector and using hash to assign partition to tasks.
